Changing profession ids to match dbc skill ids

+++ Backend/src/modules/armory/domain_value/character_info.rs
@@ -8,8 +8,8 @@ pub struct CharacterInfo {
   pub hero_class_id: u8,
   pub level: u8,
   pub gender: bool,
-  pub profession1: Option<u8>,
-  pub profession2: Option<u8>,
+  pub profession1: Option<u16>,
+  pub profession2: Option<u16>,
   pub talent_specialization: Option<String>,
   pub race_id: u8,
 }
