+++ Backend/src/modules/data/domain_value/profession.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone, Serialize, JsonSchema)]
 pub struct Profession {
-  pub id: u8,
+  pub id: u16,
   pub localization_id: u32
 }
