+++ Backend/src/modules/data/tools/profession.rs
@@ -2,12 +2,12 @@ use crate::modules::data::Data;
 use crate::modules::data::domain_value::Profession;
 
 pub trait RetrieveProfession {
-  fn get_profession(&self, id: u8) -> Option<Profession>;
+  fn get_profession(&self, id: u16) -> Option<Profession>;
   fn get_all_professions(&self) -> Vec<Profession>;
 }
 
 impl RetrieveProfession for Data {
-  fn get_profession(&self, id: u8) -> Option<Profession> {
+  fn get_profession(&self, id: u16) -> Option<Profession> {
     self.professions.get(&id)
       .and_then(|profession| Some(profession.clone()))
   }
